handle reaction messages
We could do slightly better and tell the user *which* message people
reacted to, but then we'd have to keep the uuids for each message around,
figure out how to gc it, etc.

For now, let's just render whose message the person reacted to, which
should hopefully be enough to figure out which message it was from context.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>